+++ assets/admin/components/util/content-body/content-body.jsx
@@ -5,7 +5,7 @@
  * @param {boolean} props.highlightSection Highlight this section.
  * @returns {object} The Content header.
  */
-function ContentBody({ children, id = "", highlightSection = false }) {
+function ContentBody({ id = "", highlightSection = false, children }) {
   const classes = `shadow-sm p-3 mb-3 bg-body rounded ${
     highlightSection ? "border border-warning" : ""
   }`;