+++ assets/tests/admin/data-fixtures.js
@@ -410,6 +410,80 @@ const clientConfigJson = {
   logging: false,
 };
 
+const members = [];
+Array.from(Array(31).keys()).forEach((i) => {
+  members.push({
+    "@id": "/v2/playlist-slides/01JX8KAN3R1D201EJC06AF099N",
+    "@type": "PlaylistSlide",
+    slide: {
+      "@id": "/v2/slides/00015Y0ZVC18N407JD07SM0YCF",
+      "@type": "Slide",
+    },
+    playlist: {
+      "@id": "/v2/playlists/01JQHJV7101JG70V6G0E9313PV",
+      "@type": "Playlist",
+      title: "1",
+      description:
+        "Similique voluptas aut temporibus soluta et cupiditate dolore. Ut ut totam fuga. Ullam quis ut nulla et totam neque quia.",
+      schedules: [],
+      slides: "/v2/playlists/01JQHJV7101JG70V6G0E9313PV/slides",
+      campaignScreens: [],
+      campaignScreenGroups: [],
+      tenants: [
+        {
+          "@type": "Tenant",
+          "@id": "/v2/tenants/01JDHRTN7R0JWK1KYN12WC16QX",
+        },
+      ],
+      isCampaign: false,
+      published: {
+        from: "2024-09-27T13:08:20.000Z",
+        to: "2025-05-26T09:48:41.000Z",
+      },
+      relationsChecksum: {
+        slides: "a392e0ee600c545e5604713161f14bac7460471f",
+      },
+    },
+    weight: { i },
+  });
+});
+
+const slidesPlaylist = {
+  "@context": "/contexts/PlaylistSlide",
+  "@id": "/v2/slides/01JVA5509R1V8S16PR15BB1Q4D/playlists",
+  "hydra:totalItems": 31,
+  "hydra:member": members,
+};
+
+const slideJson = {
+  "@context": "/contexts/Slide",
+  "@id": "/v2/slides/00015Y0ZVC18N407JD07SM0YCF",
+  "@type": "Slide",
+  title: "",
+  description: "",
+  templateInfo: {
+    "@id": "/v2/templates/01FP2SNGFN0BZQH03KCBXHKYHG",
+    options: [],
+  },
+  onPlaylists: [],
+  published: {
+    from: "2025-09-18T09:30:00.000Z",
+    to: null,
+  },
+  media: [],
+  content: {
+    title: "Title",
+  },
+  modifiedBy: "admin@example.com",
+  createdBy: "admin@example.com",
+  id: "01K5E2JGX3G60S73YRGR26W81J",
+  created: "2025-09-18T11:30:32.000Z",
+  modified: "2025-09-18T11:33:13.000Z",
+  relationsChecksum: {
+    templateInfo: "b2341f464f952dcd78c26990075ab5090af1286c",
+  },
+};
+
 const slidesJson1 = {
   "@id": "/v2/slides",
   "hydra:member": [
@@ -986,6 +1060,25 @@ const screensListJson = {
   ],
 };
 
+const onlyImageTextListJson = {
+  "@context": "/contexts/Template",
+  "@id": "/v2/templates",
+  "@type": "hydra:Collection",
+  "hydra:totalItems": 1,
+  "hydra:member": [
+    {
+      "@id": "/v2/templates/01FP2SNGFN0BZQH03KCBXHKYHG",
+      "@type": "Template",
+      title: "Billede og tekst",
+      modifiedBy: "",
+      createdBy: "",
+      id: "01FP2SNGFN0BZQH03KCBXHKYHG",
+      created: "2024-08-03T09:38:27.000Z",
+      modified: "2025-02-01T06:08:28.000Z",
+    },
+  ],
+};
+
 const templatesListJson = {
   "@id": "/v2/templates",
   "hydra:member": [
@@ -1076,6 +1169,15 @@ const templatesListJson = {
   "hydra:totalItems": 12,
 };
 
+const imageTextTemplate = {
+  title: "Billede og tekst",
+  modifiedBy: "",
+  createdBy: "",
+  id: "01FP2SNGFN0BZQH03KCBXHKYHG",
+  created: "2024-08-04T06:32:42.000Z",
+  modified: "2025-02-02T03:02:43.000Z",
+};
+
 const slidesListJson = {
   "@id": "/v2/slides",
   "hydra:member": [
@@ -1089,7 +1191,7 @@ const slidesListJson = {
       modifiedBy: "",
       createdBy: "",
       templateInfo: {
-        "@id": "/v2/templates/00MWCNKC4P0X5C0AT70E741E2V",
+        "@id": "/v2/templates/01FP2SNGFN0BZQH03KCBXHKYHG",
         options: [],
       },
       theme: "",
@@ -1645,4 +1747,8 @@ export {
   slidesListJson,
   themesJson,
   themesSingleJson,
+  onlyImageTextListJson,
+  slideJson,
+  imageTextTemplate,
+  slidesPlaylist,
 };